bazra/requirements.txt
amqp==5.2.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
autopep8==2.3.1
billiard==4.2.1
blinker==1.8.2
Brotli==1.1.0
CacheControl==0.14.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
ConfigArgParse==1.7
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
decorator==5.1.1
defusedxml==0.8.0rc2
distlib==0.3.8
Django==5.1.1
django-cors-headers==4.4.0
django-filter==24.3
django-prometheus==2.3.1
django-shortuuidfield==0.1.3
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-nested-routers==0.94.1
exceptiongroup==1.2.2
executing==2.1.0
fcm-django==2.2.1
filelock==3.16.1
firebase-admin==6.5.0
gevent==24.2.1
geventhttpclient==2.3.1
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.66.2
grpcio-status==1.66.2
httplib2==0.22.0
hyperlink==21.0.0
identify==2.6.1
idna==3.10
incremental==24.7.2
ipython==8.28.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
kombu==5.4.2
locust==2.31.8
Markdown==3.7
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
msgpack==1.1.0
nodeenv==1.9.1
oauthlib==3.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
pre-commit==3.8.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==5.28.2
psutil==6.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
redis==5.1.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sentry-sdk==2.15.0
service-identity==24.1.0
shortuuid==1.0.13
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
stack-data==0.6.3
swapper==1.4.0
tomli==2.0.2
traitlets==5.14.3
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
virtualenv==20.26.6
wcwidth==0.2.13
Werkzeug==3.0.4
zope.event==5.0
zope.interface==7.0.3
